% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{availability}
\alias{availability}
\title{Are measurements from given year, station, polutant available?}
\description{
Are measurements from given year, station, polutant available?
}
\author{
Mateusz Staniak
}
