% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giosDownloader.R
\docType{package}
\name{giosDownloader}
\alias{giosDownloader}
\alias{giosDownloader-package}
\title{giosDownloader: package for exploring air pollution data in Poland.}
\description{
Package provides functions that help import datasets from .xlsx files provided on GIOŚ website, 
convert them to the form suitable for analyses and calculate maxima over given period.
}
\section{Important functions}{

\code{\link{isAvailable}} checks if data are available in GIOŚ files.
\code{\link{importGiosFromXLSX}} imports data from multiple .xlsx files.
\code{\link{importGiosFromCSV}} does the same for .csv files converted from .xlsx files, which is faster.
\code{\link{calculateMaxima}} calculates maxima over given period for data important using \code{\link{importGiosFromXLSX}}.
}

\section{Datasets}{

\code{\link{availability}} Check if data are available for given station (column) and year, polutant (row).
}

