% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{station_codes}
\alias{station_codes}
\title{Codes (old & new) for stations (where pollution is measured)}
\description{
Codes (old & new) for stations (where pollution is measured)
}
\author{
Mateusz Staniak
}
